refactor: maxCapacity 변수 타입 변경

+++ src/main/java/com/goorm/tablepick/domain/reservation/service/ReservationService/ReservationServiceV0.java
@@ -52,7 +52,7 @@ public class ReservationServiceV0 {
 
         // 슬롯 카운트 검증
         Long count = reservationSlot.getCount();
-        Long maxCapacity = restaurant.getMaxCapacity();
+        int maxCapacity = restaurant.getMaxCapacity();
         if (count >= maxCapacity) {
             throw new ReservationException(ReservationErrorCode.EXCEED_RESERVATION_LIMIT);
         }
